refactor: use composition to create workflow form
